requirements.txt
pandas>=2.3.2
python-dateutil>=2.9.0
console-menu>=0.8.0